+++ eslint.config.js
@@ -24,6 +24,9 @@ export default [
         // Web Crypto API and structured clone - available in modern browsers and Electron
         crypto: 'readonly',
         structuredClone: 'readonly',
+        // Animation frame API - available in browsers and Electron
+        requestAnimationFrame: 'readonly',
+        cancelAnimationFrame: 'readonly',
         // jQuery and toastr - provided by SillyTavern host page
         $: 'readonly',
         jQuery: 'readonly',